Removed config from Service
