+++ Docker/rootfs/etc/services.d/NetDaemonApp/finish
@@ -1,4 +1,4 @@
-#!/usr/bin/execlineb -S0
+#!/usr/bin/execlineb -S1
 
 if -n { s6-test $# -ne 0 }
 if -n { s6-test ${1} -eq 256 }
