+++ src/DaemonRunner/DaemonRunner/NetDaemonExtensions.cs
@@ -1,4 +1,7 @@
 ﻿using System;
+using System.IO;
+using System.Text.Json;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
@@ -6,13 +9,20 @@ using NetDaemon.Common.Configuration;
 using NetDaemon.Daemon.Config;
 using NetDaemon.Service;
 using NetDaemon.Service.App;
+using NetDaemon.Service.Infrastructure;
+using Serilog;
+using Service.Infrastructure;
 
 namespace NetDaemon
 {
     public static class NetDaemonExtensions
     {
+        const string HassioConfigPath = "/data/options.json";
         public static IHostBuilder UseNetDaemon(this IHostBuilder hostBuilder)
         {
+            if (File.Exists(HassioConfigPath))
+                ReadHassioConfig();
+
             return hostBuilder
                 .ConfigureServices((context, services) =>
                 {
@@ -28,6 +38,21 @@ namespace NetDaemon
                     webbuilder.UseKestrel(options => { });
                     webbuilder.UseStartup<ApiStartup>();
                 });
+        }
+
+        public static IHostBuilder UseDefaultNetDaemonLogging(this IHostBuilder hostBuilder)
+        {
+            return hostBuilder
+                .ConfigureWebHostDefaults(webbuilder =>
+                {
+                    Log.Logger = SerilogConfigurator.Configure().CreateLogger();
+                    webbuilder.UseSerilog(Log.Logger);
+                });
+        }
+
+        public static void CleanupNetDaemon()
+        {
+            Log.CloseAndFlush();
         }
 
         private static void RegisterNetDaemonAssembly(IServiceCollection services)
@@ -53,6 +78,45 @@ namespace NetDaemon
                 return false;
             else
                 return true;
+        }
+
+        /// <summary>
+        ///     Reads the Home Assistant (hassio) configuration file
+        /// </summary>
+        /// <returns></returns>
+        static void ReadHassioConfig()
+        {
+            try
+            {
+                var hassAddOnSettings = JsonSerializer.Deserialize<HassioConfig>(File.ReadAllBytes(HassioConfigPath));
+
+                if (hassAddOnSettings?.LogLevel is not null)
+                    SerilogConfigurator.SetMinimumLogLevel(hassAddOnSettings.LogLevel);
+
+                if (hassAddOnSettings?.GenerateEntitiesOnStart is not null)
+                    Environment.SetEnvironmentVariable("NETDAEMON__GENERATEENTITIES", hassAddOnSettings.GenerateEntitiesOnStart.ToString());
+
+                if (hassAddOnSettings?.LogMessages is not null && hassAddOnSettings.LogMessages == true)
+                    Environment.SetEnvironmentVariable("HASSCLIENT_MSGLOGLEVEL", "Default");
+
+                _ = hassAddOnSettings?.AppSource ??
+                    throw new NullReferenceException("AppSource cannot be null");
+
+                if (hassAddOnSettings.AppSource.StartsWith("/") || hassAddOnSettings.AppSource[1] == ':')
+                {
+                    // Hard codede path
+                    Environment.SetEnvironmentVariable("NETDAEMON__APPSOURCE", hassAddOnSettings.AppSource);
+                }
+                else
+                {
+                    // We are in Hassio so hard code the path
+                    Environment.SetEnvironmentVariable("NETDAEMON__APPSOURCE", Path.Combine("/config/netdaemon", hassAddOnSettings.AppSource));
+                }
+            }
+            catch (Exception e)
+            {
+                Log.Fatal(e, "Failed to read the Home Assistant Add-on config");
+            }
         }
     }
 }
